pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "factory-env"
version = "1.0.0"
description = "Smart Factory Scheduling — OpenEnv RL Environment"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "openai>=1.0",
    "anthropic>=0.90",
    "gradio>=6.0",
    "openenv-core>=0.2.3",
    "fastapi>=0.100",
    "uvicorn>=0.23",
    "websockets>=12.0",
    "httpx>=0.27",
]

[project.scripts]
server = "server.app:main"
